[analyzer][optimizer] refactor to keep xdcontext simple
