+++ core/src/main/scala/org/apache/spark/sql/crossdata/execution/datasources/ddl.scala
@@ -19,20 +19,18 @@ import java.sql.{Date, Timestamp}
 
 import com.stratio.common.utils.components.logger.impl.SparkLoggerComponent
 import com.stratio.crossdata.connector.{TableInventory, TableManipulation}
-import com.typesafe.config.Config
 import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.expressions.Attribute
 import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan, Subquery}
 import org.apache.spark.sql.crossdata.XDContext
 import org.apache.spark.sql.crossdata.catalog.XDCatalog
-import XDCatalog._
+import org.apache.spark.sql.crossdata.catalog.XDCatalog._
 import org.apache.spark.sql.execution.RunnableCommand
 import org.apache.spark.sql.execution.datasources._
 import org.apache.spark.sql.sources.{BaseRelation, HadoopFsRelation, InsertableRelation}
 import org.apache.spark.sql.types.{StructType, _}
 
-import scala.annotation.tailrec
 import scala.language.implicitConversions
 import scala.reflect.io.File
 import scala.util.{Failure, Success, Try}
@@ -136,8 +134,8 @@ private[crossdata] case class ImportTablesUsingWithOptions(datasource: String, o
         logInfo(s"Importing table ${tableId.unquotedString}")
         val optionsWithTable = inventoryRelation.generateConnectorOpts(table, opts)
         val crossdataTable = CrossdataTable(table.tableName, table.database, table.schema, datasource, Array.empty[String], optionsWithTable)
-        import org.apache.spark.sql.crossdata.util.CreateRelationUtil._
         import XDCatalog._
+        import org.apache.spark.sql.crossdata.util.CreateRelationUtil._
         sqlContext.catalog.persistTable(crossdataTable, createLogicalRelation(sqlContext, crossdataTable))
       }
       val tableSeq = DDLUtils.tableIdentifierToSeq(tableId)
@@ -218,7 +216,7 @@ private[crossdata] case class InsertIntoTable(tableIdentifier: TableIdentifier, 
             val dataframe = convertRows(sqlContext, parsedRows, schema)
             if (sqlContext.catalog.tableHasIndex(tableIdentifier)) {
               //insert into ES because of global index
-              val crossdataIndex = sqlContext.catalog.obtainTableIndex(tableIdentifier).get
+              val crossdataIndex = sqlContext.catalog.indexMetadataByTableIdentifier(tableIdentifier).get
               val indexInsertCols = crossdataIndex.indexedCols.mkString(",") + "," + crossdataIndex.pk
 
               var values: List[Any] = Nil
